dev_requirements.txt
simba_ml==1.0.0rc15
pytorch_lightning==2.0.0
seaborn==0.12.2
torchsummary==1.5.1
pandas==2.0.2
matplotlib==3.7.1
tqdm==4.65.0
lightning==2.0.4
toml==0.10.2
pickle==4.0
Jinja2==3.1.3
MarkupSafe==2.1.5
